#90 : phase1 - standard DaemonSet, setup_flex.sh, integrate inside install, uninstall scripts and usage with flags.

+++ scripts/k8s_scbe_all_in_one/ubiquity_uninstall.sh
@@ -4,6 +4,7 @@ scripts=$(dirname $0)
 YML_DIR="./yamls"
 UBIQUITY_DB_PVC_NAME=ibm-ubiquity-db
 UTILS=$scripts/ubiquity_utils.sh
+flex_conf="ubiquity-k8s-flex.conf"
 
 [ ! -d "$YML_DIR" ] && { echo "Error: YML directory [$YML_DIR] does not exist."; exit 1; }
 [ ! -f $UTILS ] && { echo "Error: $UTILS file not found"; exit 3; }
@@ -11,8 +12,10 @@ UTILS=$scripts/ubiquity_utils.sh
 . $UTILS # include utils for wait function and status
 
 # First phase : MUST to delete the ubiquity-db deployment and ubiquity-db-pvc before deleting ubiquity and provisioner.
-kubectl delete -f $YML_DIR/ubiquity-db-deployment.yml
-sleep 30 # TODO wait till deployment stopped including the POD.
+if kubectl get deployment ubiquity-db >/dev/null 2>&1; then
+    kubectl delete -f $YML_DIR/ubiquity-db-deployment.yml
+    sleep 30 # TODO wait till deployment stopped including the POD.
+fi
 pvname=`kubectl get pvc ${UBIQUITY_DB_PVC_NAME} --no-headers -o custom-columns=name:spec.volumeName`
 kubectl delete -f ${YML_DIR}/ubiquity-db-pvc.yml
 echo "Waiting for PVC ${UBIQUITY_DB_PVC_NAME} and PV $pvname to be deleted, before delete ubiquity and provisioner."
@@ -23,6 +26,10 @@ wait_for_item_to_delete pv $pvname 10 3
 kubectl delete -f ${YML_DIR}/storage-class.yml
 kubectl delete -f $YML_DIR/ubiquity-k8s-provisioner-deployment.yml
 kubectl delete configmap k8s-config
+
+kubectl delete -f $YML_DIR/ubiquity-k8s-flex-daemonset.yml
+kubectl delete configmap $flex_conf
+
 kubectl delete -f $YML_DIR/ubiquity-deployment.yml
 kubectl delete -f $YML_DIR/ubiquity-service.yml
 kubectl delete -f $YML_DIR/ubiquity-db-service.yml
